feat(prototyper): add heap
