+++ .vscode/launch.json
@@ -1,0 +1,19 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "gdb",
+            "request": "attach",
+            "name": "Attach to gdbserver",
+            "executable": "${workspaceRoot}/target/riscv64imac-unknown-none-elf/release/rustsbi-prototyper",
+            "target": "127.0.0.1:1234",
+            "remote": true,
+            "cwd": "${workspaceRoot}/prototyper/src",
+            "valuesFormatting": "parseText",
+            "gdbpath": "/usr/bin/riscv64-linux-gnu-gdb"
+        }
+    ]
+}
