+++ prototyper/src/sbi/mod.rs
@@ -13,6 +13,7 @@ pub mod hart_context;
 pub mod logger;
 pub mod trap;
 pub mod trap_stack;
+pub mod heap;
 
 use console::{ConsoleDevice, SbiConsole};
 use hsm::SbiHsm;
