+++ prototyper/src/sbi/trap_stack.rs
@@ -5,8 +5,8 @@ use core::mem::forget;
 use fast_trap::FreeTrapStack;
 use crate::cfg::{NUM_HART_MAX, LEN_STACK_PER_HART};
 
-/// Root stack array for all harts, placed in uninitialized BSS section.
-#[link_section = ".bss.uninit"]
+/// Root stack array for all harts, placed in BSS Stack section.
+#[link_section = ".bss.stack"]
 pub(crate) static mut ROOT_STACK: [Stack; NUM_HART_MAX] = [Stack::ZERO; NUM_HART_MAX];
 
 /// Locates and initializes stack for each hart.
